Modify App major feature :
no longer need to call SET_FIO0_STATE, SET FIO1_STATE to set FI0 or FIO1

Publish a SET_FIOX_STATE = "FIO=x;VALUE=y;" with x the FIO output to set and y the value to set.

This modification was needed because other apps know FIO to be used on a config file.

So this feature is now more general.

Signed-off-by: Clément Aubry <clement.aubry@gmail.com>